requirements.txt
diffusers
transformers
torch
pillow
matplotlib